Split parser into it's own file
